dev: outsource types to separate file
